% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{dist}
\alias{dist}
\title{Distribution of data based off Categories in column}
\usage{
dist(column)
}
\description{
This functions works off categorical data. It
will return the frequency and percentage of data points
that fit to each distinct category.
}
