% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{dm_dist}
\alias{dm_dist}
\title{Distribution of data based off Categories in column}
\usage{
dist(dplyr::starwars$gender)
}
\arguments{
\item{column}{the column to run distribution checks on}
}
\description{
This functions works off categorical data. It
will return the frequency and percentage of data points
that fit to each distinct category.
}
