% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tranformations.R
\name{dm_lst2txt}
\alias{dm_lst2txt}
\title{List 2 Test}
\usage{
dm_lst2txt(column, sep = ", ")
}
\arguments{
\item{column}{the column to transform}

\item{sep}{the separator to use}
}
\description{
Takes a column that is made up of a list and converts to
concatenated values with a separator specified. Defaults to comma.
}
