% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{dm_skewness}
\alias{dm_skewness}
\title{Check the 'Skewness' of a data set}
\usage{
dm_skewness(dplyr::starwars)
}
\arguments{
\item{dataset}{the dataset to run skewness checks on}
}
\description{
This function returns the skewness of some data. The larger the skew to the
left, the stronger the negative skew. The larges the skew to the right, the stronger the
positive skew.
}
