% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{dm_types}
\alias{dm_types}
\title{Type Checking}
\usage{
types(dplyr::starwars)
}
\arguments{
\item{x}{the data to return types for}
}
\description{
This functions returns the types for a given dataset or
field.
}
