% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{dm_stddev}
\alias{dm_stddev}
\title{WIP DOESNT WORK YET
Standard Deviations of Numerical Columns}
\usage{
stdev(dplyr::starwars)
}
\description{
This functions returns the Standard Deviation
for a numerical column in a dataset.
}
